+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # Change log
 
+## VERSION
+
+* Adds `OCSInventoryWindowsAgent` [#496](https://github.com/aaronparker/evergreen/issues/496), `GoogleDrive` [#480](https://github.com/aaronparker/evergreen/issues/480), `gsudo` [#464](https://github.com/aaronparker/evergreen/issues/464)
+* Removes HDX RTME from `CitrixWorkspaceApp` [#493](https://github.com/aaronparker/evergreen/issues/493)
+* Fixes an issue with the `-Body` parameter on the internal function `Invoke-RestMethodWrapper`
+
 ## 2305.795
 
 * Fixes an issue with `MicrosoftWvdInfraAgent` where the source no longer returns the download size [#490](https://github.com/aaronparker/evergreen/issues/490)